Move TypedMoveBorrowedRustVec into vector mod
